fix(renamer): append extra number if file path already exists
This is to prevent the rewriting of existing files when bulk renaming.
It currently has a quirk: if the file exists but will be later renamed,
the suffix will be kept. For example, if two files 0.txt and 1.txt
exist, and listdir returns 1.txt first, 1.txt will end up as 0_1.txt.